1.6.0a-1.6.1 Changes
1.6.0a-1.6.0e fixes
Fixed an issue when displaying dispatch date when there are multiple
disks recoveries on the same node.
Fixed an issue with days displaying higher number then when the disk has
been failed out previously.
Fixed an issue with gen 1 internal drives not being able to be assigned
and enhanced output for gen1/2 internal disks
Fixed an issue where it was reporting garbage mds disks when the system
is setup with mixed disks
Fixed an issue where it was producing the same disk as mixed and SS
More information in the .info file for historical purposes
Added another check for disk health (current pending sectors) check
Fixed -s option for internal drives
Fixed an issue displaying last known dev path when the disk is
unavailable
1.6.0e-now
Added -x option for xfs repair on xfs mounted disks
If a disk has any current pending sectors, show_offline will mark the
disk as suspect