1.6.1a-1.6.2
Fixed an issue where if an mds unconfigured and ss configured disks are
on the same node and only shows the ss unconfigured disks.
The -a option now sorts the output list of disks found.
Added feature for inprogress disks. It will display objects recovered
per second and ETA displayed in days